chore(frontend): organize vue files better, use lazy loading
