+++ frontend/src/Main.vue
@@ -6,5 +6,5 @@
 </template>
 
 <script setup>
-import Navbar from './Navbar.vue'
+import Navbar from './components/Navbar.vue'
 </script>
