+++ frontend/tests/app-modal-dialogs.spec.js
@@ -150,6 +150,13 @@ test.describe("app modal dialogs", () => {
   }) => {
     await page.goto(`/#/open-api-ui/%2Fopenapi.yaml`);
 
+    await expect(
+      page.getByRole("heading", {
+        name: `API Documentation`,
+      }),
+      "OpenApi dialog should be visible"
+    ).toBeVisible();
+
     await page.keyboard.down("Escape");
 
     await expect(page, "should close the dialog").toHaveURL(/\/#\/$/);
@@ -159,6 +166,13 @@ test.describe("app modal dialogs", () => {
     page,
   }) => {
     await page.goto(`/#/async-api-ui/%2Fopenapi.yaml`);
+
+    await expect(
+      page.getByRole("heading", {
+        name: "AsyncAPI Documentation",
+      }),
+      "AsyncAPI dialog should be visible"
+    ).toBeVisible();
 
     await page.keyboard.down("Escape");
